calculate winner
