+++ src/App.js
@@ -1,11 +1,39 @@
-import React from "react";
+import React, { useState } from "react";
 import Board from './components/Board';
+import { calculateWinner } from "./helpers";
+
 import "./styles/root.scss"
 const App = () => {
-  return (<div className="app">
-    <h1>TIC TAC TOE </h1>
-    <Board />
-  </div>);
+
+  const [board, setBoard] = useState(Array(9).fill(null));
+  const [isXNext, setIsXNext] = useState(false);
+
+  const winner = calculateWinner(board);
+  const message = winner ? `Winner is ${winner}` : `Next Player is ${isXNext ? 'X' : 'O'}`;
+
+  const handleSquareClick = position => {
+    if (board[position] || winner) {
+      return;
+    }
+
+    setBoard(prev => {
+      return prev.map((square, pos) => {
+        if (pos === position) {
+          return isXNext ? 'X' : 'O';
+        }
+        return square;
+      });
+    });
+    setIsXNext(prev => !prev);
+  };
+
+  return (
+    <div className="app">
+      <h1>TIC TAC TOE </h1>
+      <h2>{message}</h2>
+      <Board board={board} handleSquareClick={handleSquareClick} />
+    </div>
+  );
 };
