+++ src/components/Board.js
@@ -1,28 +1,8 @@
-import React, { useState } from 'react'
+import React from 'react'
 import Square from './Square'
 
 
-
-const Board = () => {
-  const [board, setBoard] = useState(Array(9).fill(null));
-  const [isXNext, setIsXNext] = useState(false);
-
-
-  const handleSquareClick = position => {
-    if (board[position]) {
-      return;
-    }
-    console.log(board);
-    setBoard(prev => {
-      return prev.map((square, pos) => {
-        if (pos === position) {
-          return isXNext ? 'X' : 'O';
-        }
-        return square;
-      });
-    });
-    setIsXNext(prev => !prev);
-  };
+const Board = ({ board, handleSquareClick }) => {
 
   const renderSquare = position => {
     return (
